mcp_servers/weather_mcp/weather-mcp/requirements.txt
fastmcp>=2.5.0
requests>=2.28.0
python-dotenv>=1.0.0
